// frontend/src/api/generated/model/ciStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { CheckRun } from './checkRun';
import type { CiState } from './ciState';

export interface CiStatus {
  checks: CheckRun[];
  state: CiState;
  /** @minimum 0 */
  total_count: number;
}